requirements.txt
pymupdf
python-pptx
pillow
